% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarker.R
\name{findmarker}
\alias{findmarker}
\title{Marker Gene for Cell Clusters}
\usage{
findmarker(expr, individual, cluster)
}
\arguments{
\item{expr}{A numeric matrix of gene expression. Each row is a gene and each column is a cell.}

\item{individual}{A character vector indicating which individual each cell comes from, corresponding to the columns of \code{expr}.}

\item{cluster}{A numeric/character vector indicating the cell type/cluster of each cell, corresponding to the columns of \code{expr}.}
}
\value{
A list of elements. Each element correponds to one cluster and is a data.frame with six columns. The first two columns are the averaged expression of cells within or outside the cluster. The last four columns are output of \code{raisinfit}.
}
\description{
Identify marker genes for each cell cluster using RAISIN model
}
\details{
For each cell cluster, this function identifies differential genes between cells within and outside the cluster. The differential is done for all clusters
}
\examples{
res <- findmarker(expr = matrix(rnorm(100*20),nrow=100,dimnames=list(paste0('gene_',1:100),paste0('cell_',1:20))),individual=rep(c('person1','person2'),each=10),cluster=rep(c('celltype1','celltype2'),10))
}
\author{
Zhicheng Ji<zhicheng.ji@duke.edu>, Wenpin Hou, Hongkai Ji
}
