% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportiontest.R
\name{proportiontest}
\alias{proportiontest}
\title{Cell Type Proportion Testing}
\usage{
proportiontest(celltype, sample, design)
}
\arguments{
\item{celltype}{A character vector indicating which cell type each cell is.}

\item{sample}{A character vector indicating the sample that each cell comes from.}

\item{design}{A numeric matrix for sample design. Row names must be the unique values in \code{sample}. Only the second column of the design will be tested. The intercept term should be added as the first column.}
}
\value{
A data.frame with three columns: log fold change, p-value and FDR.
}
\description{
Statistical testing for cell type proportions
}
\details{
This function performs statistical testing that tests if the proportions of cell type are differential across groups, and generates corresponding p-value and FDRs.

Within each sample, the cell type proportion is calculated as the number of cells in a cell type/cluster divided by the total number of cells. The proportion is logit transformed and limma is performed to call differential.
}
\examples{
design <- cbind(1,c(1,1,0,0))
rownames(design) <- c('person1','person2','person3','person4')
res <- proportiontest(celltype=rep(c('celltype1','celltype2'),each=20),sample=rep(c('person1','person2','person3','person4'),10),design = design)
}
\author{
Zhicheng Ji<zhicheng.ji@duke.edu>, Wenpin Hou, Hongkai Ji
}
