% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raisintest.R
\name{raisintest}
\alias{raisintest}
\title{RAISIN Testing}
\usage{
raisintest(fit, coef = 2, contrast = NULL, fdrmethod = "fdr")
}
\arguments{
\item{fit}{The output from the \code{raisinfit} function.}

\item{coef}{Only useful if \code{contrast} is \code{NULL}. A numeric value indicating which coefficient is to be tested.}

\item{contrast}{A numeric vector indicating the combination of coefficients. Must have the same length as the number of columns of the fixed effect design matrix \code{X}, and the order must agree with its columns.}

\item{fdrmethod}{A character value of which method to use for multiple testing. Need to be one of the method available in \code{p.adjust}.}
}
\value{
A data.frame with four columns: (log-) fold change, statistics, p-value and FDR.
}
\description{
Statistical testing for RAISIN model
}
\details{
This function performs statistical testing that tests if a coefficient or a combination of coefficients in the design matrix of fixed effects is 0, and generates corresponding p-value and FDRs.

The test statistic follows a t-distribution. A permutation procedure is used to estimate the degree of freedom of the t-distribution. The p-value is calculated as the tail probability of the t-distribution.
}
\examples{
fit <- raisinfit(expr = matrix(rnorm(100*8),nrow=100,dimnames=list(paste0('gene_',1:100),paste0('cell_',1:8))),sample=rep(c('person1_celltype1','person2_celltype1','person1_celltype2','person3_celltype2'),each=2),testtype='paired',design = data.frame(sample=c('person1_celltype1','person2_celltype1','person1_celltype2','person3_celltype2'),individual=c('person1','person2','person1','person3'),feature=c('celltype1','celltype1','celltype2','celltype2')))
res <- raisintest(fit)
}
\author{
Zhicheng Ji<zhicheng.ji@duke.edu>, Wenpin Hou, Hongkai Ji
}
